Update README.md for new Google CA certificate
As noted in #9 and #10, the https://api.iternio.com/ API requires a new certificate to be deployed to the OVMS box. This PR proposes changes to the `README.md` file to describe the deployment of this certificate. Note that I have also removed the old text describing the deployment of the "Baltimore" certificate, as this is included in the default set of certificates on the OVMS box (ref [OVMS source code](https://github.com/openvehicles/Open-Vehicle-Monitoring-System-3/tree/fb0bdde7b420a296bdb7066f1e7eaa12ead71b67/vehicle/OVMS.V3/components/ovms_tls/trustedca)). I've confirmed that connectivity to APRB works if I have the Google cert installed and the Baltimore cert removed.